+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added example file `example_parse_from_file.dart`.
 - Changes in the file `README.md`.
 - Changes in the file `pubspec.yaml`.
+- Fixed a bug in the implementation of `VerifyGenerator`.
 
 ## 1.0.22